+++ src/main/java/technikal/task/fishmarket/repository/FishRepository.java
@@ -4,6 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import technikal.task.fishmarket.model.Fish;
 
-public interface FishRepository extends JpaRepository<Fish, Integer> {
-
-}
+public interface FishRepository extends JpaRepository<Fish, Long> { }
